test_subdomains: Add SubDomainSet mpi test.
test_subdomains: Fix mpi test.
